Move ready() implementation up to IO/ZIO level
